Remove trace_view property, use self._display
